LayoutTests/inspector/dom-debugger/event-interval-breakpoints-expected.txt

CONSOLE MESSAGE: BREAKPOINT ACTION LOG 1
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 2
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 3
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 4
Tests for Event Interval breakpoints.


== Running test suite: DOMDebugger.Event.Interval
-- Running test case: DOMDebugger.Event.Interval.AddBreakpoint
Creating "interval" Event Breakpoint...
Adding "interval" Event Breakpoint...
Firing "setInterval" on window...
PASS: Should pause before event handler is run.
CALL STACK:
0: [F] handleWindow_setInterval
ASYNC CALL STACK:
1: --- setInterval ---
2: [F] trigger_setInterval
3: [P] Global Code
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Interval.AddDisabledBreakpoint
Creating "interval" Event Breakpoint...
Adding "interval" Event Breakpoint...
Disabling "interval" Event Breakpoint...
Firing "setInterval" on window...
PASS: Should not pause for disabled breakpoint.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Interval.RemoveBreakpoint
Creating "interval" Event Breakpoint...
Adding "interval" Event Breakpoint...
Removing "interval" Event Breakpoint...
Firing "setInterval" on window...
PASS: Should not pause for removed breakpoint.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Interval.RemoveDisabledBreakpoint
Creating "interval" Event Breakpoint...
Adding "interval" Event Breakpoint...
Disabling "interval" Event Breakpoint...
Removing "interval" Event Breakpoint...
Firing "setInterval" on window...
PASS: Should not pause for removed disabled breakpoint.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Interval.RepeatFireBreakpoint
Creating "interval" Event Breakpoint...
Adding "interval" Event Breakpoint...
Firing "setInterval" on window...
PASS: Should pause before event handler is run.
CALL STACK:
0: [F] handleRepeat
ASYNC CALL STACK:
1: --- setInterval ---
2: [F] repeatSetInterval
3: [P] Global Code
PASS: Should pause before event handler is run.
CALL STACK:
0: [F] handleRepeat
ASYNC CALL STACK:
1: --- setInterval ---
2: [F] repeatSetInterval
3: [P] Global Code
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Interval.Options.Condition
Creating "interval" Event Breakpoint...
Adding "interval" Event Breakpoint...

Setting condition to 'false'...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Setting condition to 'true'...

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Interval.Options.IgnoreCount
Creating "interval" Event Breakpoint...
Adding "interval" Event Breakpoint...

Setting ignoreCount to '2'...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Interval.Options.Action.Log
Creating "interval" Event Breakpoint...
Adding "interval" Event Breakpoint...

Adding log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Interval.Options.Actions.Evaluate
Creating "interval" Event Breakpoint...
Adding "interval" Event Breakpoint...

Adding evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.
-- Running test teardown.
